feat(refreshPhpPgDoc): Serialize execution of concurrent refreshPhpPgDoc() (refs #6200)
`refreshPhpPgDoc()` now uses a master-lock to prevent concurrent
executions and uses savepoints to rollback changes in database in case
of errors.

Also:
- `workflow_init.php` and `generateDocumentClass.php` had distinct
  implementations of similar operations corresponding to
  `refreshPhpPgDoc()` from `Lib.Attr.php`. They have been rewritten to
  use the reference implementation from `Lib.Attr.php`.

+++ Class/Fdl/Class.WDoc.php
@@ -300,6 +300,9 @@ class WDoc extends Doc
             else $cid = $this->fromid;
         }
         $ordered = 1000;
+        if (($err = $this->setMasterLock(true)) !== '') {
+            return $err;
+        }
         // delete old attributes before
         $this->exec_query(sprintf("delete from docattr where docid=%d  and options ~ 'autocreated=yes'", intval($cid)));
         $this->getStates();
@@ -697,6 +700,9 @@ class WDoc extends Doc
             $oattr->labeltext = sprintf(_("%s unattach timer") , _($k));
             if ($oattr->isAffected()) $oattr->Modify();
             else $oattr->Add();
+        }
+        if (($err = $this->setMasterLock(false)) !== '') {
+            return $err;
         }
         return refreshPhpPgDoc($this->dbaccess, $cid);
     }
